Move Connections to their own files
