+++ modules/effects/arrow-effects-rx2/src/main/kotlin/arrow/effects/SingleKConnection.kt
@@ -1,0 +1,42 @@
+package arrow.effects
+
+import arrow.core.Either
+import arrow.effects.typeclasses.ExitCase
+import arrow.effects.typeclasses.MonadDefer
+
+typealias SingleKConnection = KindConnection<ForSingleK>
+typealias SingleKProc<A> = (SingleKConnection, (Either<Throwable, A>) -> Unit) -> Unit
+
+/**
+ * Connection for [SingleK].
+ *
+ * A connection is represented by a composite of `cancel` functions,
+ * [KindConnection.cancel] is idempotent and all methods are thread-safe & atomic.
+ *
+ * The cancellation functions are maintained in a stack and executed in a FIFO order.
+ *
+ * @see SingleK.async
+ */
+@Suppress("FunctionName")
+fun SingleKConnection(dummy: Unit = Unit): KindConnection<ForSingleK> = KindConnection(object : MonadDefer<ForSingleK> {
+  override fun <A> defer(fa: () -> SingleKOf<A>): SingleK<A> =
+    SingleK.defer(fa)
+
+  override fun <A> raiseError(e: Throwable): SingleK<A> =
+    SingleK.raiseError(e)
+
+  override fun <A> SingleKOf<A>.handleErrorWith(f: (Throwable) -> SingleKOf<A>): SingleK<A> =
+    fix().handleErrorWith(f)
+
+  override fun <A> just(a: A): SingleK<A> =
+    SingleK.just(a)
+
+  override fun <A, B> SingleKOf<A>.flatMap(f: (A) -> SingleKOf<B>): SingleK<B> =
+    fix().flatMap(f)
+
+  override fun <A, B> tailRecM(a: A, f: (A) -> SingleKOf<Either<A, B>>): SingleK<B> =
+    SingleK.tailRecM(a, f)
+
+  override fun <A, B> SingleKOf<A>.bracketCase(release: (A, ExitCase<Throwable>) -> SingleKOf<Unit>, use: (A) -> SingleKOf<B>): SingleK<B> =
+    fix().bracketCase(release = release, use = use)
+}) { it.value().subscribe({}, {}) }
